Some minor fixes.

+++ src/win32/AVIWrite.h
@@ -33,10 +33,10 @@ class AVIWrite {
   PAVISTREAM m_streamSound;
   AVICOMPRESSOPTIONS m_options;
   AVICOMPRESSOPTIONS *m_arrayOptions[1];
-  int m_videoFrames;
-  int m_samplesSound;
-  int m_videoFramesTotal;
-  int m_samplesSoundTotal;
+  LONG m_videoFrames;
+  LONG m_samplesSound;
+  LONG m_videoFramesTotal;
+  LONG m_samplesSoundTotal;
   LONG m_totalBytes;
   bool m_failed;
   int m_segmentNumber;
